cleanup refactoring
